+++ AccountSystemLogic/src/main/java/za/ac/nwu/as/logic/flow/FetchAccountHolderFlow.java
@@ -8,5 +8,5 @@ public interface FetchAccountHolderFlow {
 
     List<AccountHolderDto> getAllAccountHolders();
 
-    AccountHolderDto getAccountHolderById(Long memberId);
+    AccountHolderDto getAccountHolderById(int memberId);
 }
