+++ AccountSystemLogic/src/main/java/za/ac/nwu/as/logic/flow/ModifyAccountHolderFlow.java
@@ -5,9 +5,9 @@ import za.ac.nwu.as.domain.dto.AccountHolderDto;
 import java.time.LocalDate;
 
 public interface ModifyAccountHolderFlow  {
-    AccountHolderDto subtractMiles(Long memberId, int balanced);
+    AccountHolderDto subtractMiles(int memberId, int balanced);
 
     AccountHolderDto updateAccountHolder(String memberName, int balance, String currency, LocalDate startDate);
 
-    AccountHolderDto addMiles(Long memberId, int balance);
+    AccountHolderDto addMiles(int memberId, int balance);
 }
